Split XBee sensor simulator classes into sepate files
This change allows us to reuse those classes for the physical XBee
sensors. We implement an abstract base class XBee_Sensor to guide sensor
implementations.
